Update README and refactor database handling in oligo designer tools
- Simplified README instructions by removing redundant mentions of additional tools.
- Added installation command for `samtools` in the README.
- Enhanced the `ReferenceDatabase` class to ensure proper removal of index files when overwriting the database.
- Implemented a new method in `FastaParser` to create or refresh FASTA index files using `samtools`.
- Updated the `ReferenceSpecificityFilter` class to utilize the new index removal method for better file management.
- Improved cleanup process in `VCFParser` to handle compressed files and their associated index files.

+++ oligo_designer_toolsuite/database/_reference_database.py
@@ -9,7 +9,7 @@ from typing import Any, get_args
 
 from oligo_designer_toolsuite._constants import _TYPES_REF
 from oligo_designer_toolsuite._exceptions import DatabaseError
-from oligo_designer_toolsuite.utils import FastaParser, VCFParser, check_if_list
+from oligo_designer_toolsuite.utils import FastaParser, VCFParser, check_if_list, remove_index_files
 
 ############################################
 # Reference Database Class
@@ -84,7 +84,11 @@ class ReferenceDatabase:
 
         # remove all files if database should be overwritten
         if self.database_file is not None and database_overwrite:
-            os.remove(self.database_file)
+            # Remove the main database file
+            if os.path.exists(self.database_file):
+                os.remove(self.database_file)
+            remove_index_files(file_reference=self.database_file, dir_output=self.dir_output)
+
             self.database_file = None
             self.database_type = None
 
@@ -135,6 +139,8 @@ class ReferenceDatabase:
             file_ending = "fna" if self.database_type == "fasta" else "vcf.gz"
             file_database = os.path.join(dir_output, f"{filename}.{file_ending}")
             shutil.copy2(self.database_file, file_database)
+            if self.database_type == "fasta":
+                self.fasta_parser.index_fasta_file(file_fasta=file_database)
             return file_database
         else:
             raise DatabaseError("Database is empty. Nothing to be written to file.")
